coloquei o pageable no listar peddiod do gerente

+++ src/main/java/br/com/api/mgdexpress/MGD/EXPRESS/controller/masterController.java
@@ -63,8 +63,8 @@ public class masterController {
 
     @PreAuthorize("hasRole('ROLE_USER_MASTER')")
     @GetMapping("/gerente/{id}")
-    public ResponseEntity GerentesById(@PathVariable Long id){
-        return ResponseEntity.ok(gerenteRepository.getReferenceById(id));
+    public ResponseEntity GerentesById(@PathVariable Integer id){
+        return ResponseEntity.ok(gerenteRepository.getReferenceById(id.longValue()));
     }
